#CKFK20230427 Correcciones en la comunicacion

+++ app/src/main/java/com/dts/roadp/CliDet.java
@@ -366,7 +366,7 @@ public class CliDet extends PBase {
 	
 	private void showData() {
 		Cursor DT;
-		int uvis;
+		long uvis;
 		String contr,sgps="0.00000000 , 0.00000000";
 		
 		lblNom.setText("");lblRep.setText("");
@@ -388,7 +388,7 @@ public class CliDet extends PBase {
 
 			tel=DT.getString(4);
 			lblTel.setText(DT.getString(4));
-			uvis=DT.getInt(3);
+			uvis=DT.getLong(3);
 			
 			nivel=DT.getInt(6);
 			gl.nivel=nivel;
